backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.4.0
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.11.1
blinker==1.7.0
botocore==1.35.16
cachetools==5.3.2
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.3.2
cryptography==43.0.1
cycler==0.11.0
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
extra-streamlit-components==0.1.56
fastapi==0.109.1
Flask==2.2.5
fonttools==4.42.1
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.41
google-api-core==1.34.0
google-auth==2.27.0
google-cloud-appengine-logging==1.4.0
google-cloud-audit-log==0.2.5
google-cloud-core==2.4.1
google-cloud-logging==3.9.0
google-cloud-secret-manager==2.16.2
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gotrue==2.1.0
gql==3.5.0
graphql-core==3.2.3
grpc-google-iam-v1==0.12.4
grpcio==1.60.1
grpcio-status==1.48.2
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httpx==0.25.2
idna==3.7
importlib-metadata==7.0.1
importlib-resources==6.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.5
mdurl==0.1.2
multidict==6.0.5
noloco==0.6.1
numpy==1.26.4
openpyxl==3.1.5
packaging==23.2
pandas==2.3.1
pathvalidate==3.2.0
patsy==0.5.6
pdfminer.six==20221105
pdfplumber==0.7.6
pillow==10.3.0
Pint==0.23
plotly==5.14.1
plotly-express==0.4.1
pluggy==1.4.0
postgrest==0.15.0
proto-plus==1.23.0
protobuf==3.20.3
py==1.11.0
pyairtable==1.4.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydash==7.0.7
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.0.9
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.7
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==2.13.7
realtime==1.0.2
referencing==0.33.0
requests==2.32.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
scipy==1.16.1
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
st-btn-select==0.1.2
starlette==0.35.1
statsmodels==0.14.1
storage3==0.7.0
streamlit==1.37.0
streamlit-authenticator==0.2.1
StrEnum==0.4.15
supabase==2.3.6
supafunc==0.3.3
tabula==1.0.5
tabula-py==2.5.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.5.0
uvicorn==0.27.0.post1
validators==0.22.0
Wand==0.6.13
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
wordcloud==1.9.4
XlsxWriter==3.0.3
yarl==1.9.4
zipp==3.17.0
